fix trivial compiler warning. same. same. same. same. same. same. same.
2002-07-29  Charles Kerr  <charles@rebelbase.com>

* gmime-filter-strip: fix trivial compiler warning.
* gmime-message-part.c: same.
* gmime-multipart-signed.c: same.
* gmime-object.c: same.
* gmime-param.c: same.
* gmime-session.c: same.
* test-iconv.c: same.
* test-mbox.c: same.
* test-parser.c: same.
* test-partial.c: same.

+++ gmime-multipart-signed.c
@@ -411,7 +411,7 @@ g_mime_multipart_signed_verify (GMimeMultipartSigned *mps, GMimeCipherContext *c
 	valid = g_mime_cipher_verify (ctx, hash, stream, sigstream, ex);
 	
 	d(printf ("attempted to verify:\n----- BEGIN SIGNED PART -----\n%.*s----- END SIGNED PART -----\n",
-		  GMIME_STREAM_MEM (stream)->buffer->len, GMIME_STREAM_MEM (stream)->buffer->data));
+		  (int)GMIME_STREAM_MEM (stream)->buffer->len, GMIME_STREAM_MEM (stream)->buffer->data));
 	
 	g_mime_stream_unref (sigstream);
 	g_mime_stream_unref (stream);
